pkg/fileutils: move GetTotalUsedFds internal in daemon
