+++ daemon/internal/filedescriptors/filiedescriptors_linux.go
@@ -1,4 +1,4 @@
-package fileutils
+package filedescriptors
 
 import (
 	"context"
@@ -13,8 +13,6 @@ import (
 
 // GetTotalUsedFds Returns the number of used File Descriptors by
 // reading it via /proc filesystem.
-//
-// Deprecated: this function is only used internally, and will be removed in the next release.
 func GetTotalUsedFds(ctx context.Context) int {
 	ctx, span := tracing.StartSpan(ctx, "GetTotalUsedFds")
 	defer span.End()
